# Remove span metrics