--- a/samples/Samples.TracingWithoutLimits/Samples.TracingWithoutLimits.csproj
+++ b/samples/Samples.TracingWithoutLimits/Samples.TracingWithoutLimits.csproj
@@ -1,8 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <LoadManagedProfilerFromProfilerDirectory>true</LoadManagedProfilerFromProfilerDirectory>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Datadog.Trace\Datadog.Trace.csproj" />
-  </ItemGroup>
-</Project>