--- a/src/Datadog.Trace/Sampling/GlobalSamplingRule.cs
+++ b/src/Datadog.Trace/Sampling/GlobalSamplingRule.cs
@@ -31,7 +31,6 @@ namespace SignalFx.Tracing.Sampling
         public float GetSamplingRate(Span span)
         {
             Log.Debug("Using the global sampling rate: {0}", _globalRate);
-            span.SetMetric(Metrics.SamplingRuleDecision, _globalRate);
             return _globalRate;
         }
     }