--- a/src/Datadog.Trace/Sampling/RateLimiter.cs
+++ b/src/Datadog.Trace/Sampling/RateLimiter.cs
@@ -36,45 +36,35 @@ namespace SignalFx.Tracing.Sampling
 
         public bool Allowed(Span span)
         {
-            try
+            if (_maxTracesPerInterval == 0)
             {
-                if (_maxTracesPerInterval == 0)
-                {
-                    // Rate limit of 0 blocks everything
-                    return false;
-                }
+                // Rate limit of 0 blocks everything
+                return false;
+            }
 
-                if (_maxTracesPerInterval < 0)
-                {
-                    // Negative rate limit disables rate limiting
-                    return true;
-                }
-
-                WaitForRefresh();
-
-                // This must happen after the wait, because we check for window statistics, modifying this number
-                Interlocked.Increment(ref _windowChecks);
-
-                var count = _intervalQueue.Count;
-
-                if (count >= _maxTracesPerInterval)
-                {
-                    Log.Debug("Dropping trace id {0} with count of {1} for last {2}ms.", span.TraceId, count, _intervalMilliseconds);
-                    return false;
-                }
-
-                _intervalQueue.Enqueue(DateTime.Now);
-                Interlocked.Increment(ref _windowAllowed);
-
+            if (_maxTracesPerInterval < 0)
+            {
+                // Negative rate limit disables rate limiting
                 return true;
             }
-            finally
+
+            WaitForRefresh();
+
+            // This must happen after the wait, because we check for window statistics, modifying this number
+            Interlocked.Increment(ref _windowChecks);
+
+            var count = _intervalQueue.Count;
+
+            if (count >= _maxTracesPerInterval)
             {
-                // Always set the sample rate metric whether it was allowed or not
-                // DEV: Setting this allows us to properly compute metrics and debug the
-                //      various sample rates that are getting applied to this span
-                span.SetMetric(Metrics.SamplingLimitDecision, GetEffectiveRate());
+                Log.Debug("Dropping trace id {0} with count of {1} for last {2}ms.", span.TraceId, count, _intervalMilliseconds);
+                return false;
             }
+
+            _intervalQueue.Enqueue(DateTime.Now);
+            Interlocked.Increment(ref _windowAllowed);
+
+            return true;
         }
 
         public float GetEffectiveRate()