Span metrics are not used anymore but there is always at least one per span and even more with integrations/instrumentations.

Using the NLog 4.6+ no tagging benchmarks as proxies for this change, notice duration, and memory changes:

### Original

|              Method |         Categories |      Mean |     Error |    StdDev | Ratio |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|-------------------- |------------------- |----------:|----------:|----------:|------:|-------:|-------:|-------:|----------:|
|           NoTagging | NoActiveSpanSwitch |  8.191 us | 0.3922 us | 1.1127 us |  1.00 | 0.5798 | 0.2136 | 0.0916 |   3.53 KB |
|                     |                    |           |           |           |       |        |        |        |           |
| NoTaggingSpanSwitch |   ActiveSpanSwitch | 11.084 us | 0.3181 us | 0.9328 us |  1.00 | 1.1444 | 0.4883 | 0.1221 |   4.78 KB |

### Optimized

|              Method |         Categories |     Mean |     Error |    StdDev | Ratio |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|-------------------- |------------------- |---------:|----------:|----------:|------:|-------:|-------:|-------:|----------:|
|           NoTagging | NoActiveSpanSwitch | 6.443 us | 0.2315 us | 0.6752 us |  1.00 | 0.5493 | 0.2747 | 0.0305 |    3.3 KB |
|                     |                    |          |           |           |       |        |        |        |           |
| NoTaggingSpanSwitch |   ActiveSpanSwitch | 9.110 us | 0.3060 us | 0.8779 us |  1.00 | 0.8850 | 0.3967 | 0.0610 |   4.52 KB |