@pjanotti 2021-06-08
@RassK 
> Should remove references from TagsList.cs also?

TagsList exists on OTel.NET Auto, not in this repo.