@RassK 2021-06-09
> @RassK
> 
> > Should remove references from TagsList.cs also?
> 
> TagsList exists on OTel.NET Auto, not in this repo.

Correct, seems whole tagging mechanism is much simpler here.